+++ modules/inmarBidAdapter.js
@@ -1,8 +1,7 @@
-import { logError } from '../src/utils.js';
+import {logError, mergeDeep} from '../src/utils.js';
 import { config } from '../src/config.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { BANNER, VIDEO } from '../src/mediaTypes.js';
-import {getLegacyFpd} from '../src/utils/legacyFPD.js';
 
 const BIDDER_CODE = 'inmar';
 
@@ -107,5 +106,26 @@ export const spec = {
     return syncs;
   }
 };
+
+function getLegacyFpd(ortb2) {
+  if (typeof ortb2 !== 'object') return;
+
+  let duplicate = {};
+
+  Object.keys(ortb2).forEach((type) => {
+    let prop = (type === 'site') ? 'context' : type;
+    duplicate[prop] = (prop === 'context' || prop === 'user') ? Object.keys(ortb2[type]).filter(key => key !== 'data').reduce((result, key) => {
+      if (key === 'ext') {
+        mergeDeep(result, ortb2[type][key]);
+      } else {
+        mergeDeep(result, {[key]: ortb2[type][key]});
+      }
+
+      return result;
+    }, {}) : ortb2[type];
+  });
+
+  return duplicate;
+}
 
 registerBidder(spec);
